move most of palettes.hpp into palettes.cpp and expose parts in palettes.hpp. remove forward declarations in vsdata and threads in favor of #include palettes.hpp. run astyle
