+++ lib/openglvis.cpp
@@ -66,10 +66,10 @@ void Camera::GLMultRotMatrix()
    double mat[16] =
    {
       -left[0], up[0], -dir[0], 0.0,
-      -left[1], up[1], -dir[1], 0.0,
-      -left[2], up[2], -dir[2], 0.0,
-      0.0, 0.0, 0.0, 1.0
-   };
+         -left[1], up[1], -dir[1], 0.0,
+         -left[2], up[2], -dir[2], 0.0,
+         0.0, 0.0, 0.0, 1.0
+      };
 
    glMultMatrixd(mat);
 }
@@ -82,20 +82,20 @@ void Camera::GLMultTransposeRotMatrix()
    double mat_t[16] =
    {
       -left[0], -left[1], -left[2], 0.0,
-      up[0], up[1], up[2], 0.0,
-      -dir[0], -dir[1], -dir[2], 0.0,
-      0.0, 0.0, 0.0, 1.0
-   };
+         up[0], up[1], up[2], 0.0,
+         -dir[0], -dir[1], -dir[2], 0.0,
+         0.0, 0.0, 0.0, 1.0
+      };
 
    glMultMatrixd(mat_t);
 #else
    double mat[16] =
    {
       -left[0], up[0], -dir[0], 0.0,
-      -left[1], up[1], -dir[1], 0.0,
-      -left[2], up[2], -dir[2], 0.0,
-      0.0, 0.0, 0.0, 1.0
-   };
+         -left[1], up[1], -dir[1], 0.0,
+         -left[2], up[2], -dir[2], 0.0,
+         0.0, 0.0, 0.0, 1.0
+      };
 
    glMultTransposeMatrixd(mat);
 #endif
